feat: Don't give away length of saved passwords
I initially didn't do this, because I thought it might cause confusion.
Since, I've seen other services behave similarly and think it is worth
trying. An added benifit, is that this behavior allows for potentially
querying the passwords database on Linux, so that the password is only
actually retrieved when its used.